requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
pyserial>=3.5
paho-mqtt>=1.6.1
wokwi-client>=0.3
